+++ flake.nix
@@ -65,6 +65,16 @@
     "github:nix-community/nixpkgs-terraform-providers-bin";
   inputs.nixpkgs-terraform-providers-bin.inputs.nixpkgs.follows = "nixpkgs";
 
+  inputs.llm-agents.url = "github:numtide/llm-agents.nix";
+  inputs.llm-agents.inputs.nixpkgs.follows = "nixpkgs";
+  inputs.llm-agents.inputs.blueprint.inputs.systems.follows = "systems";
+  inputs.llm-agents.inputs.treefmt-nix.follows = "treefmt-nix";
+
+  inputs.neovim-nightly-overlay.url =
+    "github:nix-community/neovim-nightly-overlay";
+  inputs.neovim-nightly-overlay.inputs.flake-parts.follows = "flake-parts";
+  inputs.neovim-nightly-overlay.inputs.nixpkgs.follows = "nixpkgs";
+
   outputs = inputs:
     inputs.flake-parts.lib.mkFlake {
       inherit inputs;