+++ hosts/hyperion/home.nix
@@ -1,6 +1,6 @@
 { config, pkgs, ... }:
 let
-  platformConfigDir = if pkgs.hostPlatform.isDarwin then
+  platformConfigDir = if pkgs.stdenv.hostPlatform.isDarwin then
     "Library/Application Support"
   else
     config.xdg.configHome;