+++ modules/nextcloud.nix
@@ -17,7 +17,7 @@
       }];
 
       services.nextcloud.enable = true;
-      services.nextcloud.package = pkgs.nextcloud31;
+      services.nextcloud.package = pkgs.nextcloud32;
       services.nextcloud.hostName = hostname;
       services.nextcloud.settings.trusted_domains = [ "reflector.enzim.ee" ];
       services.nextcloud.https = true;